fingerprint: Allow devices to configure sensor location AEX Edits: Update to S
The text for enrolling a fingerprint always assumes the sensor is
located on the back.  This patch adds a config integer that defines
where the sensor is located on the device.
  0 = back
  1 = front
  2 = left side
  3 = right side

This patch also adds a new drawable that shows the front of the device
when applicable.

Change-Id: Ia8f5f8ff4ba4d13d35aed3052f60ff665dd4f294
TICKET: CYNGNOS-3089
Signed-off-by: SagarMakhar <sagarmakhar@gmail.com>

FingerprintEnrollFindSensor: Don't overlay front for side sensors

* Both the base image and the base image + overlay are equally nonsense
  for devices with a side fingerprint scanner, but for those that overlay
  a custom fingerprint enrollment graphic, this image overlay wreaks havoc
* Only set this overlay to be visible if the sensor is on the front, to
  allow devices with side sensors to only overlay the base image

Change-Id: I7cbcea4830d4526cf9f7d139e7f84f21117fa9c4
Signed-off-by: SagarMakhar <sagarmakhar@gmail.com>
Signed-off-by: Kunmun <kunmun.devroms@gmail.com>